Refs 1881: Add deeplink and routing to modals
