+++ src/Pages/ContentListTable/components/AddContent/AddContent.test.tsx
@@ -17,7 +17,10 @@ jest.mock('../../../../services/Content/ContentQueries', () => ({
   useFetchGpgKey: jest.fn(),
 }));
 
-(useFetchGpgKey as jest.Mock).mockImplementation(() => ({ fetchGpgKey: () => '' }));
+(useFetchGpgKey as jest.Mock).mockImplementation(() => ({
+  isLoading: false,
+  fetchGpgKey: () => '',
+}));
 
 (useAddContentQuery as jest.Mock).mockImplementation(() => ({
   isLoading: false,
@@ -26,12 +29,10 @@ jest.mock('../../../../services/Content/ContentQueries', () => ({
 
 jest.mock('../../../../Hooks/useDebounce', () => (value) => value);
 
-jest.mock('../../../../middleware/AppContext', () => ({
-  useAppContext: () => ({ rbac: { read: true, write: true } }),
-}));
+jest.mock('../../../../Hooks/useNotification', () => () => ({ notify: () => null }));
 
-jest.mock('../../../../services/Notifications/Notifications', () => ({
-  useNotification: () => ({ notify: () => null }),
+jest.mock('react-router-dom', () => ({
+  useNavigate: jest.fn(),
 }));
 
 const passingValidationMetaDataSigNotPresent = [
@@ -40,49 +41,6 @@ const passingValidationMetaDataSigNotPresent = [
     url: { ...passingValidationErrorData[0], metadata_signature_present: false },
   },
 ];
-
-it('expect AddContent button to render and be disabled', async () => {
-  (useValidateContentList as jest.Mock).mockImplementation(() => ({
-    isLoading: false,
-    mutateAsync: async () => defaultValidationErrorData,
-  }));
-
-  const { queryByText } = render(
-    <ReactQueryTestWrapper>
-      <AddContent isDisabled />
-    </ReactQueryTestWrapper>,
-  );
-
-  const button = queryByText('Add repositories');
-  await waitFor(() => {
-    expect(button).toBeInTheDocument();
-    expect(button).toHaveAttribute('disabled');
-  });
-});
-
-it('expect AddContent modal to open/close successfully', async () => {
-  (useValidateContentList as jest.Mock).mockImplementation(() => ({
-    isLoading: false,
-    mutateAsync: async () => defaultValidationErrorData,
-  }));
-
-  const { queryByText } = render(
-    <ReactQueryTestWrapper>
-      <AddContent />
-    </ReactQueryTestWrapper>,
-  );
-
-  const button = queryByText('Add repositories');
-  expect(button).toBeInTheDocument();
-  await act(async () => {
-    button?.click();
-  });
-  expect(queryByText('Add custom repositories')).toBeInTheDocument();
-  await act(async () => {
-    queryByText('Cancel')?.click();
-  });
-  expect(queryByText('Add custom repositories')).not.toBeInTheDocument();
-});
 
 it('expect "name" input to show a validation error', async () => {
   (useValidateContentList as jest.Mock).mockImplementation(() => ({
@@ -96,20 +54,16 @@ it('expect "name" input to show a validation error', async () => {
     </ReactQueryTestWrapper>,
   );
 
-  const button = queryByText('Add repositories');
-  expect(button).toBeInTheDocument();
-  await act(async () => {
-    button?.click();
-  });
   const nameInput = queryByPlaceholderText('Enter name');
   expect(nameInput).toBeInTheDocument();
   if (nameInput) {
     await act(async () => {
       fireEvent.change(nameInput, { target: { value: 'b' } });
-      fireEvent.blur(nameInput);
     });
   }
-  expect(queryByText('Too Short!')).toBeInTheDocument();
+  waitFor(() => {
+    expect(queryByText('Too Short!')).toBeInTheDocument();
+  });
 });
 
 it('expect "url" input to show a validation error', async () => {
@@ -124,20 +78,16 @@ it('expect "url" input to show a validation error', async () => {
     </ReactQueryTestWrapper>,
   );
 
-  const button = queryByText('Add repositories');
-  expect(button).toBeInTheDocument();
-  await act(async () => {
-    button?.click();
-  });
   const urlInput = queryByPlaceholderText('https://');
   expect(urlInput).toBeInTheDocument();
   if (urlInput) {
     await act(async () => {
       fireEvent.change(urlInput, { target: { value: 'bobTheBuilder' } });
-      fireEvent.blur(urlInput);
     });
   }
-  expect(queryByText('Invalid URL')).toBeInTheDocument();
+  waitFor(() => {
+    expect(queryByText('Invalid URL')).toBeInTheDocument();
+  });
 });
 
 it('expect "Package and metadata verification" to be pre-selected', async () => {
@@ -153,32 +103,24 @@ it('expect "Package and metadata verification" to be pre-selected', async () => 
     </ReactQueryTestWrapper>,
   );
 
-  const button = queryByText('Add repositories');
-  expect(button).toBeInTheDocument();
-  await act(async () => {
-    button?.click();
-  });
   const urlInput = queryByPlaceholderText('https://');
   expect(urlInput).toBeInTheDocument();
-  if (urlInput) {
-    await act(async () => {
-      fireEvent.change(urlInput, { target: { value: 'https://bobTheBuilder.com' } });
-      fireEvent.blur(urlInput);
-    });
-  }
-  expect(queryByText('Invalid URL')).not.toBeInTheDocument();
+  await act(async () => {
+    fireEvent.change(urlInput as HTMLElement, { target: { value: 'https://bobTheBuilder.com' } });
+  });
+  waitFor(() => {
+    expect(queryByText('Invalid URL')).not.toBeInTheDocument();
+  });
+
   const gpgKeyInput = queryByPlaceholderText('Paste GPG key or URL here');
   expect(gpgKeyInput).toBeInTheDocument();
-  if (gpgKeyInput) {
-    await act(async () => {
-      fireEvent.change(gpgKeyInput, { target: { value: 'aRealGPGKey' } });
-    });
-  }
+  await act(async () => {
+    fireEvent.change(gpgKeyInput as HTMLElement, { target: { value: 'aRealGPGKey' } });
+  });
 
-  expect(queryByText('Package and metadata verification')).toBeInTheDocument();
-  const packageMetaDataInput = queryByLabelText('Package and metadata verification');
-
-  expect(packageMetaDataInput).toHaveAttribute('checked');
+  waitFor(() => {
+    expect(queryByLabelText('Package and metadata verification')).toHaveAttribute('checked');
+  });
 });
 
 it('expect "Package verification only" to be pre-selected', async () => {
@@ -194,30 +136,22 @@ it('expect "Package verification only" to be pre-selected', async () => {
     </ReactQueryTestWrapper>,
   );
 
-  const button = queryByText('Add repositories');
-  expect(button).toBeInTheDocument();
-  await act(async () => {
-    button?.click();
-  });
   const urlInput = queryByPlaceholderText('https://');
   expect(urlInput).toBeInTheDocument();
-  if (urlInput) {
-    await act(async () => {
-      fireEvent.change(urlInput, { target: { value: 'https://bobTheBuilder.com' } });
-    });
-  }
+
+  await act(async () => {
+    fireEvent.change(urlInput as HTMLElement, { target: { value: 'https://bobTheBuilder.com' } });
+  });
   expect(queryByText('Invalid URL')).not.toBeInTheDocument();
   const gpgKeyInput = queryByPlaceholderText('Paste GPG key or URL here');
   expect(gpgKeyInput).toBeInTheDocument();
-  if (gpgKeyInput) {
-    await act(async () => {
-      fireEvent.change(gpgKeyInput, { target: { value: 'aRealGPGKey' } });
-      fireEvent.blur(gpgKeyInput);
-    });
-  }
-  expect(queryByText('Package verification only')).toBeInTheDocument();
-  const packageVerificationOnly = queryByLabelText('Package verification only');
-  expect(packageVerificationOnly).toHaveAttribute('checked');
+  await act(async () => {
+    fireEvent.change(gpgKeyInput as HTMLElement, { target: { value: 'aRealGPGKey' } });
+  });
+  waitFor(() => {
+    expect(queryByText('Package verification only')).toBeInTheDocument();
+    expect(queryByLabelText('Package verification only')).toHaveAttribute('checked');
+  });
 });
 
 it('Add content', async () => {
@@ -233,12 +167,6 @@ it('Add content', async () => {
     </ReactQueryTestWrapper>,
   );
 
-  const button = queryByText('Add repositories');
-  expect(button).toBeInTheDocument();
-  await act(async () => {
-    button?.click();
-  });
-
   const nameInput = queryByPlaceholderText('Enter name');
   expect(nameInput).toBeInTheDocument();
   const urlInput = queryByPlaceholderText('https://');
@@ -246,22 +174,27 @@ it('Add content', async () => {
   const gpgKeyInput = queryByPlaceholderText('Paste GPG key or URL here');
   if (urlInput && nameInput && gpgKeyInput) {
     await act(async () => {
-      fireEvent.change(urlInput, { target: { value: 'https://google.com/' } });
-    });
-    await act(async () => {
       fireEvent.change(nameInput, { target: { value: 'superCoolName' } });
     });
+
+    waitFor(() => {
+      expect(nameInput).toHaveAttribute('value', 'superCoolName');
+    });
+
+    await act(async () => {
+      fireEvent.change(urlInput, { target: { value: 'https://google.com/' } });
+    });
+
+    waitFor(() => {
+      expect(urlInput?.getAttribute('value')).toBe('https://google.com/');
+    });
+
     await act(async () => {
       fireEvent.change(gpgKeyInput, { target: { value: 'test GPG key' } });
     });
-    await act(async () => {
-      fireEvent.blur(urlInput);
-    });
-    await act(async () => {
-      fireEvent.blur(nameInput);
-    });
-    await act(async () => {
-      fireEvent.blur(gpgKeyInput);
+
+    waitFor(() => {
+      expect(gpgKeyInput?.getAttribute('value')).toBe('test GPG key');
     });
   }
 
@@ -270,25 +203,37 @@ it('Add content', async () => {
 
   expect(queryByText('Invalid URL')).not.toBeInTheDocument();
   const addAnotherButton = queryByText('Add another repository');
-  expect(addAnotherButton?.getAttribute('disabled')).toBeNull();
+  waitFor(() => {
+    expect(addAnotherButton?.getAttribute('aria-disabled')).toBe('false');
+  });
   if (addAnotherButton) {
     await act(async () => {
       fireEvent.click(addAnotherButton);
     });
   }
   const secondRemoveButton = queryAllByText('Remove')[1];
-  expect(secondRemoveButton).toBeInTheDocument();
+
+  waitFor(() => {
+    expect(secondRemoveButton).toBeInTheDocument();
+  });
+
   if (secondRemoveButton) {
     await act(async () => {
       fireEvent.click(secondRemoveButton);
     });
   }
+
   const saveButton = queryByText('Save');
-  expect(saveButton?.getAttribute('disabled')).toBeNull();
+  waitFor(() => {
+    expect(saveButton?.getAttribute('disabled')).toBeNull();
+  });
+
   if (saveButton) {
     await act(async () => {
       fireEvent.click(saveButton);
     });
   }
-  expect(queryByText('Add custom repository')).not.toBeInTheDocument();
+  waitFor(() => {
+    expect(queryByText('Add custom repository')).not.toBeInTheDocument();
+  });
 });
