+++ src/Pages/ContentListTable/components/EditContentModal/EditContentModal.test.tsx
@@ -24,6 +24,7 @@ jest.mock('../../../../services/Content/ContentQueries', () => ({
   useEditContentQuery: () => ({ isLoading: false }),
   useValidateContentList: jest.fn(),
   useFetchGpgKey: () => ({ fetchGpgKey: () => '' }),
+  useFetchContent: () => ({ data: editItem, isError: false }),
 }));
 
 jest.mock('react-query', () => ({
@@ -31,12 +32,16 @@ jest.mock('react-query', () => ({
   useQueryClient: jest.fn(),
 }));
 
-jest.mock('../../../../services/Notifications/Notifications', () => ({
-  useNotification: () => ({ notify: () => null }),
-}));
-
+jest.mock('../../../../Hooks/useNotification', () => () => ({ notify: () => null }));
 jest.mock('../../../../Hooks/useDebounce', () => (value) => value);
 jest.mock('../../../../middleware/AppContext', () => ({ useAppContext: () => ({}) }));
+
+jest.mock('react-router-dom', () => ({
+  useNavigate: jest.fn(),
+  useLocation: () => ({
+    search: 'repoUUIDS=someLongUUID',
+  }),
+}));
 
 it('Open, confirming values, edit an item, enabling Save button', async () => {
   (useValidateContentList as jest.Mock).mockImplementation(() => ({
@@ -49,16 +54,12 @@ it('Open, confirming values, edit an item, enabling Save button', async () => {
 
   const { queryByText, queryByPlaceholderText, queryAllByLabelText } = render(
     <ReactQueryTestWrapper>
-      <EditContentModal open values={[editItem]} setClosed={() => undefined} />
+      <EditContentModal />
     </ReactQueryTestWrapper>,
   );
 
-  expect(
-    queryByText('Edit by completing the form. Default values may be provided automatically.'),
-  ).toBeInTheDocument();
-  const NameTextfield = queryByPlaceholderText('Enter name');
-  expect(NameTextfield).toBeInTheDocument();
-  expect(NameTextfield).toHaveAttribute('value', editItem.name);
+  expect(queryByPlaceholderText('Enter name')).toBeInTheDocument();
+  expect(queryByPlaceholderText('Enter name')).toHaveAttribute('value', editItem.name);
 
   const UrlTextfield = queryByPlaceholderText('https://');
   expect(UrlTextfield).toBeInTheDocument();
