+++ src/Pages/ContentListTable/components/PackageModal/PackageModal.tsx
@@ -31,11 +31,13 @@ import { useEffect, useMemo, useState } from 'react';
 import { createUseStyles } from 'react-jss';
 import { SkeletonTable } from '@redhat-cloud-services/frontend-components';
 import Hide from '../../../../components/Hide/Hide';
-import { ContentItem, PackageItem } from '../../../../services/Content/ContentApi';
+import { PackageItem } from '../../../../services/Content/ContentApi';
 import { useGetPackagesQuery } from '../../../../services/Content/ContentQueries';
 import { SearchIcon } from '@patternfly/react-icons';
 import useDebounce from '../../../../Hooks/useDebounce';
 import EmptyPackageState from './components/EmptyPackageState';
+import { useNavigate, useParams } from 'react-router-dom';
+import { BASE_ROUTE } from '../../../../Routes/useTabbedRoutes';
 
 const useStyles = createUseStyles({
   description: {
@@ -71,18 +73,13 @@ const useStyles = createUseStyles({
   },
 });
 
-interface Props {
-  rowData: ContentItem;
-  closeModal: () => void;
-}
-
 const perPageKey = 'packagePerPage';
 
-export default function PackageModal({
-  rowData: { name, uuid, package_count: packageCount },
-  closeModal,
-}: Props) {
+export default function PackageModal() {
   const classes = useStyles();
+  const { repoUUID: uuid } = useParams();
+
+  const navigate = useNavigate();
   const storedPerPage = Number(localStorage.getItem(perPageKey)) || 20;
   const [page, setPage] = useState(1);
   const [perPage, setPerPage] = useState(storedPerPage);
@@ -108,10 +105,15 @@ export default function PackageModal({
   const {
     isLoading,
     isFetching,
-    error,
     isError,
     data = { data: [], meta: { count: 0, limit: 20, offset: 0 } },
-  } = useGetPackagesQuery(uuid, packageCount, page, perPage, debouncedSearchQuery, sortString);
+  } = useGetPackagesQuery(uuid as string, page, perPage, debouncedSearchQuery, sortString);
+
+  useEffect(() => {
+    if (isError) {
+      onClose();
+    }
+  }, [isError]);
 
   const onSetPage: OnSetPage = (_, newPage) => setPage(newPage);
 
@@ -139,8 +141,7 @@ export default function PackageModal({
     };
   };
 
-  // Error is caught in the wrapper component
-  if (isError) throw error;
+  const onClose = () => navigate(BASE_ROUTE);
 
   const {
     data: packageList = [],
@@ -161,15 +162,11 @@ export default function PackageModal({
       ouiaSafe={fetchingOrLoading}
       variant={ModalVariant.medium}
       title='Packages'
-      description={
-        <p className={classes.description}>
-          View list of packages for <b>{name}</b>
-        </p>
-      }
+      description={<p className={classes.description}>View list of packages</p>}
       isOpen
-      onClose={closeModal}
+      onClose={onClose}
       footer={
-        <Button key='close' variant='secondary' onClick={closeModal}>
+        <Button key='close' variant='secondary' onClick={onClose}>
           Close
         </Button>
       }
