+++ src/components/EmptyTableState/EmptyTableState.test.tsx
@@ -3,9 +3,8 @@ import { ReactQueryTestWrapper } from '../../testingHelpers';
 import EmptyTableState from './EmptyTableState';
 import { Button } from '@patternfly/react-core';
 
-jest.mock('../../services/Notifications/Notifications', () => ({
-  useNotification: () => ({ notify: () => null }),
-}));
+jest.mock('../../Hooks/useNotification', () => () => ({ notify: () => null }));
+
 jest.mock('../../middleware/AppContext', () => ({
   useAppContext: () => ({}),
 }));
