+++ src/services/Features/FeatureQueries.ts
@@ -1,25 +1,19 @@
-import { AlertVariant } from '@patternfly/react-core';
 import { useState } from 'react';
-import { useNotification } from '../Notifications/Notifications';
 import { Features, getFeatures } from './FeatureApi';
+import useErrorNotification from '../../Hooks/useErrorNotification';
 
 export const useFetchFeaturesQuery = () => {
   const [isLoading, setIsLoading] = useState(false);
-  const { notify } = useNotification();
+  const errorNotifier = useErrorNotification();
 
   const fetchFeatures = async (): Promise<Features | null> => {
     setIsLoading(true);
     let features: Features | null = null;
     try {
       features = await getFeatures();
-    } catch ({ response = {} }: any) {
-      const { data } = response as { data: { message: string | undefined } | string };
-      const description = typeof data === 'string' ? data : data?.message;
-      notify({
-        variant: AlertVariant.danger,
-        title: 'Error fetching features',
-        description,
-      });
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    } catch (err: any) {
+      errorNotifier('Error fetching features', 'An error occurred', err);
     }
     setIsLoading(false);
     return features;
