+++ src/testingHelpers.tsx
@@ -128,6 +128,7 @@ export const defaultContentItem: ContentItem = {
   last_introspection_time: '2023-03-07 17:13:48.619192 -0500 EST',
   failed_introspections_count: 0,
   metadata_verification: false,
+  snapshot: false,
 };
 
 export const defaultPopularRepository: PopularRepository = {
